move files in com/daml
