+++ daml-lf/transaction-test-lib/src/main/scala/com/daml/lf/value/test/ValueGenerators.scala
@@ -3,6 +3,7 @@
 
 package com.daml.lf
 package value
+package test
 
 import com.daml.lf.data.Ref._
 import com.daml.lf.data._
